+++ HelpTechService/Attention/Interfaces/REST/Resources/Job/AssignJobDetailResource.cs
@@ -2,6 +2,6 @@
 {
     public record AssignJobDetailResource
         (int Id, DateTime WorkDate,
-        decimal Time, double LaborBudget,
-        double MaterialBudget);
+        decimal Time, decimal LaborBudget,
+        decimal MaterialBudget);
 }
